+++ front/app/[id]/_services/user.service.ts
@@ -7,11 +7,11 @@ export class UserService {
   constructor(private userRepository: UserRepository) {
     this.userRepository = UserRepository.getInstance()
   }
-  async findByPublicId(id: string,token:Token): Promise<User | null> {
+  async findByPublicId(id: string, token: Token): Promise<User | null> {
     try {
       // front で無理やり型を合わせているため、パフォーマンスが良くないかも
-      const DBUser = await this.userRepository.findByPublicId(id,token)
-      if(!DBUser) return null
+      const DBUser = await this.userRepository.findByPublicId(id, token)
+      if (!DBUser) return null
       const user: User = {
         ...DBUser,
         posts: DBUser?.posts.map((post) => ({
@@ -21,7 +21,7 @@ export class UserService {
             nickname: DBUser.nickname,
             img_url: DBUser.img_url,
             isPublic: DBUser.isPublic,
-            publicId:id
+            publicId: id,
           },
         })),
       }
