+++ front/components/shared/PostInputDrawer.tsx
@@ -1,0 +1,8 @@
+import { auth } from '@clerk/nextjs'
+import PostInput from './PostInput'
+
+export default async function PostInputDrawer() {
+  const { getToken } = auth()
+  const token = await getToken()
+  return <PostInput token={token} />
+}
